Add scroll-spy-active and scroll-spy-link directives
This allow to handle menu active class and links easily.

It also add supports for multiple scroll-spy directives in the same app with data-scroll-spy-id, sets the value to last index when scrolling to the bottom, and fix issues related to computation of offset when scroll-spy container is not positionned at top: 0.

Dependencies have been upgraded, ESLint added, webpack configuration enhanced.